requirements.txt
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
